# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vyaya API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "common.validation_failed",
              "common.not_found",
              "common.invalid_cursor",
              "common.internal",
              "common.dependency_unavailable",
              "auth.unauthenticated",
              "txn.already_reversed",
              "txn.transfer_metadata_requires_group",
              "category.parent_kind_mismatch",
              "asset.invalid_valuation_sign",
              "import.invalid_file",
              "import.already_committed",
              "import.invalid_state",
              "recurring.no_occurrences"
            ]
          },
          "reqId": {
            "type": "string"
          },
          "timestamp": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "retryable": {
            "type": "boolean"
          },
          "errors": {
            "type": ["array", "null"],
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": ["path", "code", "message"]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "reqId",
          "timestamp",
          "retryable",
          "errors"
        ]
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "categoryId": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "type": {
            "type": "string",
            "enum": ["expense", "income"]
          },
          "amountMinor": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "occurredAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40
            },
            "maxItems": 20,
            "default": []
          },
          "id": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "currency": {
            "type": "string",
            "enum": ["INR"]
          },
          "source": {
            "type": "string",
            "enum": ["manual", "csv_import", "recurring", "api"]
          },
          "status": {
            "type": "string",
            "enum": ["posted", "reversed", "reversal"]
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid"
          },
          "reversalOf": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "reversedBy": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "transferGroupId": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "createdAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "updatedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": [
          "accountId",
          "type",
          "amountMinor",
          "occurredAt",
          "description",
          "id",
          "userId",
          "currency",
          "source",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "Transfer": {
        "type": "object",
        "properties": {
          "transferGroupId": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "fromTransaction": {
            "type": "object",
            "properties": {
              "accountId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "categoryId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "type": {
                "type": "string",
                "enum": ["expense", "income"]
              },
              "amountMinor": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "occurredAt": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "description": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 40
                },
                "maxItems": 20,
                "default": []
              },
              "id": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "userId": {
                "type": "string",
                "minLength": 1
              },
              "currency": {
                "type": "string",
                "enum": ["INR"]
              },
              "source": {
                "type": "string",
                "enum": ["manual", "csv_import", "recurring", "api"]
              },
              "status": {
                "type": "string",
                "enum": ["posted", "reversed", "reversal"]
              },
              "idempotencyKey": {
                "type": "string",
                "format": "uuid"
              },
              "reversalOf": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "reversedBy": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "transferGroupId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "createdAt": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "updatedAt": {
                "type": ["string", "null"],
                "format": "date-time"
              }
            },
            "required": [
              "accountId",
              "type",
              "amountMinor",
              "occurredAt",
              "description",
              "id",
              "userId",
              "currency",
              "source",
              "status",
              "createdAt",
              "updatedAt"
            ]
          },
          "toTransaction": {
            "type": "object",
            "properties": {
              "accountId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "categoryId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "type": {
                "type": "string",
                "enum": ["expense", "income"]
              },
              "amountMinor": {
                "type": "integer",
                "minimum": 1,
                "maximum": 9007199254740991
              },
              "occurredAt": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "description": {
                "type": "string",
                "minLength": 1,
                "maxLength": 500
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 40
                },
                "maxItems": 20,
                "default": []
              },
              "id": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "userId": {
                "type": "string",
                "minLength": 1
              },
              "currency": {
                "type": "string",
                "enum": ["INR"]
              },
              "source": {
                "type": "string",
                "enum": ["manual", "csv_import", "recurring", "api"]
              },
              "status": {
                "type": "string",
                "enum": ["posted", "reversed", "reversal"]
              },
              "idempotencyKey": {
                "type": "string",
                "format": "uuid"
              },
              "reversalOf": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "reversedBy": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "transferGroupId": {
                "type": "string",
                "pattern": "^[a-f\\d]{24}$/i"
              },
              "createdAt": {
                "type": ["string", "null"],
                "format": "date-time"
              },
              "updatedAt": {
                "type": ["string", "null"],
                "format": "date-time"
              }
            },
            "required": [
              "accountId",
              "type",
              "amountMinor",
              "occurredAt",
              "description",
              "id",
              "userId",
              "currency",
              "source",
              "status",
              "createdAt",
              "updatedAt"
            ]
          }
        },
        "required": ["transferGroupId", "fromTransaction", "toTransaction"]
      },
      "TransferReversal": {
        "type": "object",
        "properties": {
          "transferGroupId": {
            "type": "string",
            "pattern": "^[a-f\\d]{24}$/i"
          },
          "legs": {
            "type": "array",
            "prefixItems": [
              {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "categoryId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["expense", "income"]
                  },
                  "amountMinor": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "occurredAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 40
                    },
                    "maxItems": 20,
                    "default": []
                  },
                  "id": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "userId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "currency": {
                    "type": "string",
                    "enum": ["INR"]
                  },
                  "source": {
                    "type": "string",
                    "enum": ["manual", "csv_import", "recurring", "api"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["posted", "reversed", "reversal"]
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reversalOf": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "reversedBy": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "transferGroupId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "createdAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  }
                },
                "required": [
                  "accountId",
                  "type",
                  "amountMinor",
                  "occurredAt",
                  "description",
                  "id",
                  "userId",
                  "currency",
                  "source",
                  "status",
                  "createdAt",
                  "updatedAt"
                ]
              },
              {
                "type": "object",
                "properties": {
                  "accountId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "categoryId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["expense", "income"]
                  },
                  "amountMinor": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "occurredAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 40
                    },
                    "maxItems": 20,
                    "default": []
                  },
                  "id": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "userId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "currency": {
                    "type": "string",
                    "enum": ["INR"]
                  },
                  "source": {
                    "type": "string",
                    "enum": ["manual", "csv_import", "recurring", "api"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["posted", "reversed", "reversal"]
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reversalOf": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "reversedBy": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "transferGroupId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "createdAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  }
                },
                "required": [
                  "accountId",
                  "type",
                  "amountMinor",
                  "occurredAt",
                  "description",
                  "id",
                  "userId",
                  "currency",
                  "source",
                  "status",
                  "createdAt",
                  "updatedAt"
                ]
              }
            ]
          }
        },
        "required": ["transferGroupId", "legs"]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/transactions/{transactionId}/reverse": {
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$/i"
            },
            "required": true,
            "name": "transactionId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Reversal, or natural replay for the already-reversed transaction",
            "headers": {
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": ["true"]
                },
                "required": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Already reversed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers": {
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fromAccountId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "toAccountId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$/i"
                  },
                  "amountMinor": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "occurredAt": {
                    "type": ["string", "null"],
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 40
                    },
                    "maxItems": 20,
                    "default": []
                  }
                },
                "required": [
                  "fromAccountId",
                  "toAccountId",
                  "amountMinor",
                  "occurredAt",
                  "description"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "201": {
            "description": "Created transfer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transfer"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transfers/{transferGroupId}/reverse": {
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$/i"
            },
            "required": true,
            "name": "transferGroupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Group reversal, or natural replay keyed by the original transfer group",
            "headers": {
              "Idempotency-Replayed": {
                "schema": {
                  "type": "string",
                  "enum": ["true"]
                },
                "required": false
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReversal"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}